version inestable

+++ unstable/ss-admin/limpiando.php
@@ -4,17 +4,19 @@ include_once '../script/php/functions.php';
 conectar();
 
 $matricula=trim($_POST['matricula']);
+$todo=$_POST['todo'];
 
-$query="SELECT CriterioAlu FROM `alumno_ss_fca` WHERE `MatriculaAlu`='$matricula' LIMIT 1;";
-$result=mysql_query($query) or die(mysql_error());
-$rows=mysql_fetch_array($result);
-$criterio=$rows['CriterioAlu'];
-
-if(!(empty($criterio))){
-	$query="DELETE FROM `evaluacion_$criterio` WHERE `MatriculaAlu`='$matricula' LIMIT 1;";
-	$result=mysql_query($query) or die(mysql_error());	
+if($todo){
+	$query="SELECT CriterioAlu FROM `alumno_ss_fca` WHERE `MatriculaAlu`='$matricula' LIMIT 1;";
+	$result=mysql_query($query) or die(mysql_error());
+	$rows=mysql_fetch_array($result);
+	$criterio=$rows['CriterioAlu'];
+	
+	if(!(empty($criterio))){
+		$query="DELETE FROM `evaluacion_$criterio` WHERE `MatriculaAlu`='$matricula' LIMIT 1;";
+		$result=mysql_query($query) or die(mysql_error());	
+	}
 }
-
 
 
 $query="DELETE FROM `alumno_ss_fca` WHERE `MatriculaAlu`='$matricula' LIMIT 1;";
